Options for Papaya Image

Options for Papaya Image

papayaOptions(
  alpha = NULL,
  icon = NULL,
  interpolation = NULL,
  lut = NULL,
  negative_lut = NULL,
  max = NULL,
  maxPercent = NULL,
  min = NULL,
  minPercent = NULL,
  rotation = NULL,
  rotationPoint = NULL,
  parametric = NULL,
  symmetric = NULL
)

validateOptions(...)

Arguments

alpha

The overlay image alpha level (0 to 1).

icon

Image url to use as the toolbar icon (rescaled to 18x18).

interpolation

If true, image will have a smooth display (default true).

lut

The color table name (built-in or custom).

negative_lut

The color table name (built-in or custom) used by the negative side of the parametric pair.

max

The display range maximum.

maxPercent

The display range maximum as a percentage of image max.

min

The display range minimum.

minPercent

The display range minimum as a percentage of image max.

rotation

An array of rotation degrees about each of the three axes (e.g., [0,0,45] = 45 degree rotation about Z axis)

rotationPoint

The point of rotation. Options are "center" (default), "origin", or "crosshairs".

parametric

If true, loads two views of the same image, one in the positive and one in the negative range. (Also see combineParametric.)

symmetric

If true, sets the negative range of a parametric pair to the same size as the positive range.

...

additional arguments (not used)

Examples

papayaOptions(alpha = 0.5)
#> $alpha
#> [1] 0.5
#> 
papayaOptions()
#> NULL
papayaOptions(rotation = c(0, 0, 45),
parametric = TRUE, rotationPoint = "center")
#> $rotation
#> [1]  0  0 45
#> 
#> $rotationPoint
#> [1] "center"
#> 
#> $parametric
#> [1] TRUE
#>