Shiny bindings for papayaWidget

Output and render functions for using papayaWidget within Shiny applications and interactive Rmd documents.

papayaOutput(outputId, width = "100%", height = "400px")

renderPapaya(expr, env = parent.frame(), quoted = FALSE)

Arguments

outputId
output variable to read from

width, height
Must be a valid CSS unit (like '100%', '400px', 'auto') or a number, which will be coerced to a string and have 'px' appended.

expr
An expression that generates a papayaWidget

env
The environment in which to evaluate expr.

quoted
Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.

Examples

file = system.file("examples", "kirby21.R", package = "papayaWidget")
if (requireNamespace("shiny", quietly = TRUE) &&
    file.exists(file)) {
   app = source(file)
}
papayaOutput("random_id")
#> <div id="random_id" style="width:100%; height:400px; " class="papayaWidget html-widget html-widget-output"></div>
renderPapaya({papaya()})
#> <shiny.render.function>